fix: delete commits and rename files
