+++ index.ts
@@ -6,6 +6,6 @@ const HTTP_PORT = 8181;
 console.log(`Start static http server on the ${HTTP_PORT} port!`);
 httpServer.listen(HTTP_PORT);
 
+WSServer.on('listening', onListen);
 WSServer.on('connection', onConnect);
 WSServer.on('close', onClose);
-WSServer.on('listening', onListen);
