+++ src/data/userData.ts
@@ -1,8 +1,3 @@
 import { IRegUser } from '../interface/interface.js';
 
-export const userList: IRegUser[] = [
-  {
-    name: '12345',
-    password: '12345',
-  },
-];
+export const userList: Array<IRegUser> = [];
