+++ src/router/router.ts
@@ -3,11 +3,9 @@ import { IAddShips, IAttack, IInstruction, IRegUser, UserAddToRoom } from '../in
 import { sendRegResponse, sendUpdateWinnersToAll } from '../controller/controller.js';
 import { addUserToRoom, createNewRoom, sendUpdateRoomStateToAll } from '../controller/roomController.js';
 import { addShipsToGameBoard } from '../controller/shipController.js';
-import { handleAttack } from '../controller/gameCotroller.js';
+import { handleAttack } from '../controller/gameController.js';
 
 export const router = <T>(ws: WebSocket, data: IInstruction<T>) => {
-  console.log('Data in router', data);
-
   switch (data.type) {
     case 'reg':
       sendRegResponse(ws, data as IInstruction<IRegUser>);
@@ -23,14 +21,14 @@ export const router = <T>(ws: WebSocket, data: IInstruction<T>) => {
       addUserToRoom(ws, data as IInstruction<UserAddToRoom>);
       console.log('add user to room');
       break;
+    case 'add_ships':
+      addShipsToGameBoard(ws, data as IInstruction<IAddShips>);
+      console.log('add ships');
+      break;
     case 'attack':
     case 'randomAttack':
       handleAttack(data as IInstruction<IAttack>);
       console.log('random attack');
-      break;
-    case 'add_ships':
-      addShipsToGameBoard(ws, data as IInstruction<IAddShips>);
-      console.log('add ships');
       break;
     default:
       console.log('Invalid input');
